# backend/app/api/actividades.py
# date: usado para validar que la fecha de la actividad quede dentro del
# rango de fechas del evento al que pertenece.
from datetime import date

# APIRouter/Depends/HTTPException/status: ver detalle en app/api/auth.py.
from fastapi import APIRouter, Depends, HTTPException, status
# select: construcción declarativa de consultas SQL.
from sqlalchemy import select
from sqlalchemy.orm import Session

# Dependencia que exige que el usuario autenticado tenga un rol específico.
from app.api.users import require_system_role
from app.db import get_db
from app.models import Actividad, CategoriaActividad, Conferencista, Evento, Usuario
from app.schemas import ActividadCreate, ActividadResponse

# Router con el prefijo "/api/v1/actividades", agrupado bajo el tag
# "Actividades" (agenda de un evento: charlas, talleres, paneles). Cubre la
# creación de actividades, la agenda del evento y la asociación/
# desasociación de conferencistas (HU06); el CRUD completo de actividades
# queda fuera de este alcance.
router = APIRouter(prefix="/api/v1/actividades", tags=["Actividades"])
# Router separado con el prefijo "/api/v1/eventos": expone la creación de
# una actividad anidada bajo su evento (POST /api/v1/eventos/{evento_id}/
# actividades), a diferencia del resto de los endpoints de este módulo, que
# operan sobre una actividad ya creada identificada por su propio id.
evento_actividades_router = APIRouter(prefix="/api/v1/eventos", tags=["Actividades"])
# Dependencia reutilizada en los endpoints de escritura: exige que el
# usuario autenticado tenga el rol de sistema "organizador" (leer/listar
# queda público, igual que en app/api/eventos.py y app/api/conferencistas.py).
organizer_required = Depends(require_system_role("organizador"))


# Busca un evento por id y lanza 404 si no existe, para no crear una
# actividad apuntando a un evento inexistente.
def _get_evento_or_404(db: Session, id_evento: int) -> Evento:
    evento = db.get(Evento, id_evento)
    if evento is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Evento no encontrado")
    return evento


# Busca una categoría de actividad activa por id y lanza 404 si no existe o
# está desactivada, para no clasificar una actividad con una categoría que
# ya no se ofrece.
def _get_categoria_activa_or_404(db: Session, id_categoria: int) -> CategoriaActividad:
    categoria = db.get(CategoriaActividad, id_categoria)
    if categoria is None or not categoria.activo:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Categoría de actividad no encontrada")
    return categoria


# Busca una actividad por id y lanza 404 si no existe.
def _get_actividad_or_404(db: Session, actividad_id: int) -> Actividad:
    actividad = db.get(Actividad, actividad_id)
    if actividad is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Actividad no encontrada")
    return actividad


# Busca un conferencista por id y lanza 404 si no existe.
def _get_conferencista_or_404(db: Session, conferencista_id: int) -> Conferencista:
    conferencista = db.get(Conferencista, conferencista_id)
    if conferencista is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Conferencista no encontrado")
    return conferencista


# Valida que la fecha de la actividad quede dentro del rango de fechas del
# evento (fecha_inicio <= fecha <= fecha_fin). La coherencia entre
# hora_inicio y hora_fin de la propia actividad ya la valida ActividadCreate
# a nivel de schema; esta validación adicional necesita el evento cargado
# desde la base de datos, por lo que se hace en el endpoint.
def _validate_fecha_dentro_evento(evento: Evento, fecha: date) -> None:
    if not (evento.fecha_inicio <= fecha <= evento.fecha_fin):
        raise HTTPException(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            detail="La fecha de la actividad debe estar dentro del rango de fechas del evento",
        )


# Crea una nueva actividad dentro de un evento, validando que el evento y la
# categoría existan y que la fecha de la actividad esté dentro del rango de
# fechas del evento.
@evento_actividades_router.post(
    "/{evento_id}/actividades", response_model=ActividadResponse, status_code=status.HTTP_201_CREATED
)
def create_actividad(
    evento_id: int,
    data: ActividadCreate,
    db: Session = Depends(get_db),
    _: Usuario = organizer_required,
) -> ActividadResponse:
    evento = _get_evento_or_404(db, evento_id)
    categoria = _get_categoria_activa_or_404(db, data.id_categoria)
    _validate_fecha_dentro_evento(evento, data.fecha)

    actividad = Actividad(
        id_evento=evento.id_evento,
        id_categoria=categoria.id_categoria,
        titulo=data.titulo.strip(),
        descripcion=data.descripcion,
        fecha=data.fecha,
        hora_inicio=data.hora_inicio,
        hora_fin=data.hora_fin,
        lugar=data.lugar,
        modalidad=data.modalidad,
        cupo=data.cupo,
    )
    db.add(actividad)
    db.commit()
    db.refresh(actividad)
    return ActividadResponse.model_validate(actividad)


# Obtiene una actividad puntual por id, incluyendo los conferencistas
# actualmente asociados. Público, igual que el resto de los "get" de
# recursos de agenda.
@router.get("/{actividad_id}", response_model=ActividadResponse)
def get_actividad(actividad_id: int, db: Session = Depends(get_db)) -> ActividadResponse:
    actividad = _get_actividad_or_404(db, actividad_id)
    return ActividadResponse.model_validate(actividad)


# Lista la agenda de un evento (todas sus actividades), ordenada por fecha
# y hora de inicio. Pública, igual que el resto de los "get" de recursos de
# agenda.
@evento_actividades_router.get("/{evento_id}/agenda", response_model=list[ActividadResponse])
def get_agenda_evento(evento_id: int, db: Session = Depends(get_db)) -> list[ActividadResponse]:
    _get_evento_or_404(db, evento_id)

    query = (
        select(Actividad)
        .where(Actividad.id_evento == evento_id)
        .order_by(Actividad.fecha, Actividad.hora_inicio)
    )
    actividades = db.scalars(query).all()
    return [ActividadResponse.model_validate(actividad) for actividad in actividades]


# Asocia un conferencista a una actividad. Valida que ambos existan y que el
# conferencista no esté ya asociado (409 Conflict si ya lo está, para evitar
# asociaciones duplicadas silenciosas).
@router.post(
    "/{actividad_id}/conferencistas/{conferencista_id}",
    response_model=ActividadResponse,
    status_code=status.HTTP_201_CREATED,
)
def associate_conferencista(
    actividad_id: int,
    conferencista_id: int,
    db: Session = Depends(get_db),
    _: Usuario = organizer_required,
) -> ActividadResponse:
    actividad = _get_actividad_or_404(db, actividad_id)
    conferencista = _get_conferencista_or_404(db, conferencista_id)

    if conferencista in actividad.conferencistas:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail="El conferencista ya está asociado a esta actividad",
        )

    actividad.conferencistas.append(conferencista)
    db.commit()
    db.refresh(actividad)
    return ActividadResponse.model_validate(actividad)


# Desasocia un conferencista de una actividad. Valida que ambos existan y
# que la asociación exista (404 si el conferencista no estaba asociado a
# esta actividad).
@router.delete("/{actividad_id}/conferencistas/{conferencista_id}", status_code=status.HTTP_204_NO_CONTENT)
def disassociate_conferencista(
    actividad_id: int,
    conferencista_id: int,
    db: Session = Depends(get_db),
    _: Usuario = organizer_required,
) -> None:
    actividad = _get_actividad_or_404(db, actividad_id)
    conferencista = _get_conferencista_or_404(db, conferencista_id)

    if conferencista not in actividad.conferencistas:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="El conferencista no está asociado a esta actividad",
        )

    actividad.conferencistas.remove(conferencista)
    db.commit()