# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eventia API",
    "description": "\nAPI REST del sistema **Eventia** para la gestión de eventos académicos:\neventos, actividades (agenda), conferencistas, inscripciones y\nparametrización de roles.\n\n### Cómo autenticarse\n1. Ejecutar `POST /api/v1/auth/login` con email y contraseña.\n2. Copiar el `access_token` de la respuesta.\n3. Presionar **Authorize** e ingresar el token (sin el prefijo `Bearer`).\n\n### Permisos por rol\n* **administrador**: parametrización de roles (`/api/v1/roles/...`).\n* **organizador**: alta/edición de eventos, actividades y conferencistas.\n* Los endpoints de consulta (GET públicos), el login y las inscripciones no\n  requieren token.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/eventos": {
      "post": {
        "tags": [
          "Eventos"
        ],
        "summary": "Create Evento",
        "operationId": "create_evento_api_v1_eventos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "List Eventos",
        "operationId": "list_eventos_api_v1_eventos_get",
        "parameters": [
          {
            "name": "id_tipo_evento",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id Tipo Evento"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "fecha_desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Desde"
            }
          },
          {
            "name": "fecha_hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha Hasta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoResponse"
                  },
                  "title": "Response List Eventos Api V1 Eventos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/eventos/{evento_id}": {
      "get": {
        "tags": [
          "Eventos"
        ],
        "summary": "Get Evento",
        "operationId": "get_evento_api_v1_eventos__evento_id__get",
        "parameters": [
          {
            "name": "evento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evento Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Eventos"
        ],
        "summary": "Update Evento",
        "operationId": "update_evento_api_v1_eventos__evento_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evento Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/actividades/{actividad_id}": {
      "get": {
        "tags": [
          "Actividades"
        ],
        "summary": "Get Actividad",
        "operationId": "get_actividad_api_v1_actividades__actividad_id__get",
        "parameters": [
          {
            "name": "actividad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actividad Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActividadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/actividades/{actividad_id}/conferencistas/{conferencista_id}": {
      "post": {
        "tags": [
          "Actividades"
        ],
        "summary": "Associate Conferencista",
        "operationId": "associate_conferencista_api_v1_actividades__actividad_id__conferencistas__conferencista_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "actividad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actividad Id"
            }
          },
          {
            "name": "conferencista_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conferencista Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActividadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Actividades"
        ],
        "summary": "Disassociate Conferencista",
        "operationId": "disassociate_conferencista_api_v1_actividades__actividad_id__conferencistas__conferencista_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "actividad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Actividad Id"
            }
          },
          {
            "name": "conferencista_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conferencista Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eventos/{evento_id}/actividades": {
      "post": {
        "tags": [
          "Actividades"
        ],
        "summary": "Create Actividad",
        "operationId": "create_actividad_api_v1_eventos__evento_id__actividades_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "evento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evento Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActividadCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActividadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eventos/{evento_id}/agenda": {
      "get": {
        "tags": [
          "Actividades"
        ],
        "summary": "Get Agenda Evento",
        "operationId": "get_agenda_evento_api_v1_eventos__evento_id__agenda_get",
        "parameters": [
          {
            "name": "evento_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Evento Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActividadResponse"
                  },
                  "title": "Response Get Agenda Evento Api V1 Eventos  Evento Id  Agenda Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActividadCreate": {
        "properties": {
          "id_categoria": {
            "type": "integer",
            "title": "Id Categoria"
          },
          "titulo": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Titulo"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "hora_inicio": {
            "type": "string",
            "format": "time",
            "title": "Hora Inicio"
          },
          "hora_fin": {
            "type": "string",
            "format": "time",
            "title": "Hora Fin"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "modalidad": {
            "type": "string",
            "enum": [
              "PRESENCIAL",
              "VIRTUAL",
              "HIBRIDA"
            ],
            "title": "Modalidad",
            "default": "PRESENCIAL"
          },
          "cupo": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cupo"
          }
        },
        "type": "object",
        "required": [
          "id_categoria",
          "titulo",
          "fecha",
          "hora_inicio",
          "hora_fin",
          "cupo"
        ],
        "title": "ActividadCreate"
      },
      "ActividadResponse": {
        "properties": {
          "id_actividad": {
            "type": "integer",
            "title": "Id Actividad"
          },
          "id_evento": {
            "type": "integer",
            "title": "Id Evento"
          },
          "id_categoria": {
            "type": "integer",
            "title": "Id Categoria"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "hora_inicio": {
            "type": "string",
            "format": "time",
            "title": "Hora Inicio"
          },
          "hora_fin": {
            "type": "string",
            "format": "time",
            "title": "Hora Fin"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "modalidad": {
            "type": "string",
            "title": "Modalidad"
          },
          "cupo": {
            "type": "integer",
            "title": "Cupo"
          },
          "conferencistas": {
            "items": {
              "$ref": "#/components/schemas/ConferencistaResponse"
            },
            "type": "array",
            "title": "Conferencistas",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id_actividad",
          "id_evento",
          "id_categoria",
          "titulo",
          "descripcion",
          "fecha",
          "hora_inicio",
          "hora_fin",
          "lugar",
          "modalidad",
          "cupo"
        ],
        "title": "ActividadResponse"
      },
...
      "EventoCreate": {
        "properties": {
          "id_tipo_evento": {
            "type": "integer",
            "title": "Id Tipo Evento"
          },
          "nombre": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "cupo_maximo": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Cupo Maximo"
          },
          "estado": {
            "type": "string",
            "enum": [
              "borrador",
              "publicado"
            ],
            "title": "Estado",
            "default": "borrador"
          },
          "politica": {
            "$ref": "#/components/schemas/PoliticaInscripcionData"
          }
        },
        "type": "object",
        "required": [
          "id_tipo_evento",
          "nombre",
          "fecha_inicio",
          "fecha_fin",
          "cupo_maximo"
        ],
        "title": "EventoCreate"
      },
      "EventoResponse": {
        "properties": {
          "id_evento": {
            "type": "integer",
            "title": "Id Evento"
          },
          "id_tipo_evento": {
            "type": "integer",
            "title": "Id Tipo Evento"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "type": "string",
            "format": "date",
            "title": "Fecha Fin"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "cupo_maximo": {
            "type": "integer",
            "title": "Cupo Maximo"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "politica": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PoliticaInscripcionData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id_evento",
          "id_tipo_evento",
          "nombre",
          "descripcion",
          "fecha_inicio",
          "fecha_fin",
          "lugar",
          "cupo_maximo",
          "estado"
        ],
        "title": "EventoResponse"
      },
      "EventoUpdate": {
        "properties": {
          "id_tipo_evento": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Tipo Evento"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "fecha_inicio": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Inicio"
          },
          "fecha_fin": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Fin"
          },
          "lugar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Lugar"
          },
          "cupo_maximo": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cupo Maximo"
          },
          "estado": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "borrador",
                  "publicado"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado"
          },
          "politica": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PoliticaInscripcionData"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "EventoUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PoliticaInscripcionData": {
        "properties": {
          "fecha_limite": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Limite"
          },
          "requiere_aprobacion": {
            "type": "boolean",
            "title": "Requiere Aprobacion",
            "default": false
          },
          "permite_lista_espera": {
            "type": "boolean",
            "title": "Permite Lista Espera",
            "default": false
          },
          "min_asistencia_certificado": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Min Asistencia Certificado",
            "default": 75
          }
        },
        "type": "object",
        "title": "PoliticaInscripcionData"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Mismo servidor que sirve la documentación"
    },
    {
      "url": "http://localhost:8000",
      "description": "Backend local (uvicorn)"
    }
  ]
}